fix(vm): remove json tags from embedded structs in vm model
When running the generator it would not concider the json tags and would
create a flat structure without a surrounding spec objec in the
typescript model. However the json structure would have this spec
object. To make sure all the models are alligned we remove the json tags
from all embedded objects.